Apply PR review changes
see also: https://github.com/YetAnotherKeyboard/firmware/pull/1
